requirements.txt
asgiref==3.4.1
autopep8==1.5.7
azure-ai-textanalytics==5.1.0
azure-common==1.1.27
azure-core==1.16.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
fastapi==0.66.0
google-api-core==1.31.0
google-auth==1.32.1
googleapis-common-protos==1.53.0
gunicorn==20.0.4
h11==0.12.0
httptools==0.2.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
opencensus-ext-azure==1.0.8
packaging==21.0
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.8.2
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
